split payment module fixtures
