+++ saleor/payment/tests/fixtures/__init__.py
@@ -1,0 +1,6 @@
+from .gateway_config import *  # noqa: F403
+from .gateway_response import *  # noqa: F403
+from .payment import *  # noqa: F403
+from .payment_data import *  # noqa: F403
+from .transaction_event import *  # noqa: F403
+from .transaction_item import *  # noqa: F403
