+++ saleor/payment/tests/fixtures/gateway_config.py
@@ -1,0 +1,13 @@
+import pytest
+
+from ....payment.interface import GatewayConfig
+
+
+@pytest.fixture
+def dummy_gateway_config():
+    return GatewayConfig(
+        gateway_name="Dummy",
+        auto_capture=True,
+        supported_currencies="USD",
+        connection_params={"secret-key": "nobodylikesspanishinqusition"},
+    )
